+++ http4k-contract/src/main/kotlin/org/http4k/contract/NoRenderer.kt
@@ -9,7 +9,7 @@ import org.http4k.lens.Failure
 object NoRenderer : ContractRenderer {
     override fun description(contractRoot: PathSegments, security: Security, routes: List<ContractRoute>) = Response(OK)
 
-    override fun badRequest(failures: List<Failure>): Response = Response(BAD_REQUEST)
+    override fun badRequest(failures: List<Failure>) = Response(BAD_REQUEST)
 
     override fun notFound(): Response = Response(NOT_FOUND)
 }