+++ http4k-contract/src/test/kotlin/org/http4k/contract/ContractRoutingHttpHandlerTest.kt
@@ -3,40 +3,52 @@ package org.http4k.contract
 import com.natpryce.hamkrest.and
 import com.natpryce.hamkrest.assertion.assertThat
 import com.natpryce.hamkrest.equalTo
+import org.http4k.core.Body
 import org.http4k.core.Filter
 import org.http4k.core.Method.GET
 import org.http4k.core.Method.OPTIONS
 import org.http4k.core.Request
 import org.http4k.core.Response
+import org.http4k.core.Status.Companion.BAD_REQUEST
 import org.http4k.core.Status.Companion.NOT_IMPLEMENTED
 import org.http4k.core.Status.Companion.OK
 import org.http4k.core.Status.Companion.UNAUTHORIZED
 import org.http4k.core.UriTemplate
 import org.http4k.core.then
 import org.http4k.core.with
-import org.http4k.format.Argo
+import org.http4k.format.Jackson
+import org.http4k.format.Jackson.auto
 import org.http4k.hamkrest.hasBody
 import org.http4k.hamkrest.hasStatus
 import org.http4k.lens.Header
 import org.http4k.lens.Path
 import org.http4k.lens.Query
+import org.http4k.lens.int
 import org.http4k.routing.RoutedResponse
-import org.http4k.routing.RoutingHttpHandler
 import org.http4k.routing.RoutingHttpHandlerContract
 import org.http4k.routing.bind
 import org.http4k.routing.routes
 import org.junit.jupiter.api.Test
 
 class ContractRoutingHttpHandlerTest : RoutingHttpHandlerContract() {
-    override val handler: RoutingHttpHandler =
-        contract(SimpleJson(Argo), "/", validPath bindContract GET to { Response(OK).with(header of header(it)) })
+
+    private data class ARandomObject(val field: String)
+
+    override val handler =
+        contract(SimpleJson(Jackson), "/",
+            validPath bindContract GET to { Response(OK).with(header of header(it)) },
+            "/bad-request" bindContract GET to { Query.int().required("foo")(it); Response(OK) },
+            "/bad-request-body" bindContract GET to { Body.auto<ARandomObject>().toLens()(it); Response(OK) }
+        )
 
     private val header = Header.optional("FILTER")
+
+    override val expectedNotFoundBody = "{\"message\":\"No route found on this path. Have you used the correct HTTP verb?\"}"
 
     @Test
     fun `by default the description lives at the route`() {
         Request(GET, "/root")
-        val response = ("/root" bind contract(SimpleJson(Argo), security = ApiKey(Query.required("goo"), { false }))).invoke(Request(GET, "/root"))
+        val response = ("/root" bind contract(SimpleJson(Jackson), security = ApiKey(Query.required("goo"), { false }))).invoke(Request(GET, "/root"))
         assertThat(response.status, equalTo(OK))
         assertThat(response.bodyString(), equalTo("""{"resources":{}}"""))
     }
@@ -47,7 +59,7 @@ class ContractRoutingHttpHandlerTest : RoutingHttpHandlerContract() {
             { next(it.with(header of "true")) }
         }
 
-        val root = "/root" bind contract(SimpleJson(Argo), "/docs",
+        val root = "/root" bind contract(SimpleJson(Jackson), "/docs",
             "/" bindContract GET to { Response(OK).with(header of header(it)) })
         val withRoute = filter.then(root)
 
@@ -105,7 +117,7 @@ class ContractRoutingHttpHandlerTest : RoutingHttpHandlerContract() {
 
     @Test
     fun `applies security and responds with a 401 to unauthorized requests`() {
-        val root = "/root" bind contract(SimpleJson(Argo), "", ApiKey(Query.required("key"), { it == "bob" }),
+        val root = "/root" bind contract(SimpleJson(Jackson), "", ApiKey(Query.required("key"), { it == "bob" }),
             "/bob" bindContract GET to { Response(OK) }
         )
 
@@ -115,7 +127,7 @@ class ContractRoutingHttpHandlerTest : RoutingHttpHandlerContract() {
 
     @Test
     fun `pre-security filter is applied before security`() {
-        val root = "/root" bind contract(SimpleJson(Argo), "", ApiKey(Query.required("key"), { it == "bob" }),
+        val root = "/root" bind contract(SimpleJson(Jackson), "", ApiKey(Query.required("key"), { it == "bob" }),
             "/bob" bindContract GET to { Response(OK) }
         ).withFilter(Filter { next ->
             {
@@ -128,7 +140,7 @@ class ContractRoutingHttpHandlerTest : RoutingHttpHandlerContract() {
 
     @Test
     fun `post-security filter is applied after security`() {
-        val root = "/root" bind contract(SimpleJson(Argo), "", ApiKey(Query.required("key"), { it == "bob" }),
+        val root = "/root" bind contract(SimpleJson(Jackson), "", ApiKey(Query.required("key"), { it == "bob" }),
             "/bob" bindContract GET to { Response(OK).body(it.body) }
         ).withPostSecurityFilter(Filter { next ->
             {
@@ -141,7 +153,7 @@ class ContractRoutingHttpHandlerTest : RoutingHttpHandlerContract() {
 
     @Test
     fun `applies security and responds with a 200 to authorized requests`() {
-        val root = "/root" bind contract(SimpleJson(Argo), "", ApiKey(Query.required("key"), { it == "bob" }),
+        val root = "/root" bind contract(SimpleJson(Jackson), "", ApiKey(Query.required("key"), { it == "bob" }),
             "/bob" bindContract GET to { Response(OK) }
         )
 
@@ -151,7 +163,7 @@ class ContractRoutingHttpHandlerTest : RoutingHttpHandlerContract() {
 
     @Test
     fun `can change path to description route`() {
-        val response = ("/root/foo" bind contract(SimpleJson(Argo), "/docs/swagger.json"))
+        val response = ("/root/foo" bind contract(SimpleJson(Jackson), "/docs/swagger.json"))
             .invoke(Request(GET, "/root/foo/docs/swagger.json"))
         assertThat(response.status, equalTo(OK))
     }
@@ -171,5 +183,19 @@ class ContractRoutingHttpHandlerTest : RoutingHttpHandlerContract() {
         val response = (filter.then(contract))(Request(GET, "/test"))
         assertThat(response.status, equalTo(OK))
         assertThat(response.bodyString(), equalTo("[bar]"))
+    }
+
+    @Test
+    fun `handles bad request - parameter`() {
+        assertThat(handler(Request(GET, "/bad-request")),
+            hasStatus(BAD_REQUEST) and
+                hasBody("""{"message":"Missing/invalid parameters","params":[{"name":"foo","type":"query","datatype":"integer","required":true,"reason":"Missing"}]}"""))
+    }
+
+    @Test
+    fun `handles bad request - body`() {
+        assertThat(handler(Request(GET, "/bad-request-body")),
+            hasStatus(BAD_REQUEST) and
+                hasBody("""{"message":"Missing/invalid parameters","params":[{"name":"body","type":"body","datatype":"object","required":true,"reason":"Invalid"}]}"""))
     }
 }