+++ http4k-contract/src/test/kotlin/org/http4k/contract/NoRendererTest.kt
@@ -3,23 +3,24 @@ package org.http4k.contract
 import com.natpryce.hamkrest.assertion.assertThat
 import com.natpryce.hamkrest.equalTo
 import org.http4k.core.Response
-import org.http4k.core.Status
+import org.http4k.core.Status.Companion.BAD_REQUEST
+import org.http4k.core.Status.Companion.NOT_FOUND
+import org.http4k.core.Status.Companion.OK
 import org.junit.jupiter.api.Test
 
 class NoRendererTest {
     @Test
     fun `renders not found`() {
-        assertThat(NoRenderer.notFound(), equalTo(Response(Status.Companion.NOT_FOUND)))
+        assertThat(NoRenderer.notFound(), equalTo(Response(NOT_FOUND)))
     }
 
     @Test
     fun `renders bad request`() {
-        assertThat(NoRenderer.badRequest(listOf()), equalTo(Response(Status.Companion.BAD_REQUEST)))
+        assertThat(NoRenderer.badRequest(listOf()), equalTo(Response(BAD_REQUEST)))
     }
 
     @Test
     fun `renders description`() {
-        assertThat(NoRenderer.description(Root, NoSecurity, listOf()), equalTo(Response(Status.Companion.OK)))
+        assertThat(NoRenderer.description(Root, NoSecurity, listOf()), equalTo(Response(OK)))
     }
-
 }