+++ http4k-contract/src/test/kotlin/org/http4k/contract/OpenApiTest.kt
@@ -1,5 +1,5 @@
 package org.http4k.contract
 
-import org.http4k.format.Argo
+import org.http4k.format.Jackson
 
-class OpenApiTest : ContractRendererContract(OpenApi(ApiInfo("title", "1.2", "module description"), Argo))
+class OpenApiTest : ContractRendererContract(OpenApi(ApiInfo("title", "1.2", "module description"), Jackson))